ODROID-COMMON: common: add proxy function for U-boot environments
This patch is to add a proxy functions to manage U-boot environments in multiple
boot media. The functions, saveenv() and env_init(), are defined in the multiple
files 'common/env_*.c' and they cannot be compiled together since their names
are duplicated. Therefore, the proxy functions will arbitrate the funections as
per booting media type such as MMC or SPI.

Change-Id: I247764e1e6fb1b9daaf1990d57b5a60b50dbbbab
Signed-off-by: Dongjin Kim <tobetter@gmail.com>