ODROID-COMMON: usb: add to perform USB scanning silently
This patch is to allow traverse the USB ports without scannig the attached
devices in order to supply the initial power to the connected devices. The
oridignal implementation is to scan the USB devices and this gives some time
to USB hub until power is stable. But this procedure is not necessary only
when supplying the power on boot and later USB devices can be scanned.

This hack is required to supply a power to a certain device that is not
acutually required to USB transmission but needs a power supply.

Change-Id: I0e8a786ddc29a843bd2da560e6ada8e5382e1ea9
Signed-off-by: Dongjin Kim <tobetter@gmail.com>